apps/xfzauth/serializers.py
from rest_framework import serializers
from .models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('pk', 'telephone', 'username', 'email', 'is_active', 'is_staff')
