requirements.txt
aliyun-python-sdk-core-v3==2.11.0
aliyun-python-sdk-dysmsapi==1.0.0
asn1crypto==0.24.0
backcall==0.1.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.1
decorator==4.3.0
Django==2.1.3
django-debug-toolbar==1.10.1
django-haystack==2.8.1
django-shortuuidfield==0.1.3
djangorestframework==3.9.0
idna==2.7
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
jieba==0.39
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pudb==2018.1
pycparser==2.19
pycryptodome==3.7.0
Pygments==2.2.0
PyMySQL==0.9.2
python-memcached==1.59
pytz==2018.7
qiniu==7.2.2
requests==2.20.1
shortuuid==0.5.0
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
urllib3==1.24.1
urwid==2.0.1
wcwidth==0.1.7
Whoosh==2.7.4
